requirements.txt
torch>=2.5.0
torchvision>=0.20.0
pytorch-fid>=0.3.0
scipy>=1.17.0
matplotlib>=3.10
numpy>=2.0
einops>=0.8
tqdm>=4.67
pyyaml>=6.0
